quotetutorial/requirements.txt
attrs==19.3.0
Automat==0.7.0
cffi==1.13.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
mysql-connector-python==8.0.18
parsel==1.5.2
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.9.0
pyOpenSSL==19.0.0
queuelib==1.5.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
Twisted==19.7.0
w3lib==1.21.0
zope.interface==4.6.0
